Guava issue 1322: deal with a potential overflow issue in ImmutableTable. ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=46160775

+++ guava/src/com/google/common/collect/RegularImmutableTable.java
@@ -172,7 +172,7 @@ abstract class RegularImmutableTable<R, C, V> extends ImmutableTable<R, C, V> {
 
     // use a dense table if more than half of the cells have values
     // TODO(gak): tune this condition based on empirical evidence
-    return (cellList.size() > ((rowSpace.size() * columnSpace.size()) / 2)) ?
+    return (cellList.size() > (((long) rowSpace.size() * columnSpace.size()) / 2)) ?
         new DenseImmutableTable<R, C, V>(cellList, rowSpace, columnSpace) :
         new SparseImmutableTable<R, C, V>(cellList, rowSpace, columnSpace);
   }
